+++ src/app/tests/unit/assets/lazy-model-list-test.js
@@ -239,6 +239,14 @@ lazyModelListSuite.add(new Y.Test.Case({
         Assert.areSame('baz', models[1].get('foo'));
     },
 
+    'revive() should preserve clientIds': function () {
+        this.list.add({foo: 'bar', clientId: 'model-foo'});
+
+        var model = this.list.revive(0);
+
+        Assert.areSame('model-foo', model.get('clientId'), 'revived model should have the same clientId as the original object');
+    },
+
     '_isInList() should indicate whether an item is in the list': function () {
         var item = {foo: 'bar'};